Added 1281 Digits product and sum difference x1.5
PROBLEM: Given an integer number n, return the difference between the product of its digits and the sum of its digits. Note test case where sum > product.
https://leetcode.com/problems/subtract-the-product-and-sum-of-digits-of-an-integer/

SOLUTION: Using closures to convert i32 -> String, iterate by character, map each character to a digit, and sum them, then repeat for the product and return the difference. There is no absolute value: input of n=12 is expected to return -1 (which is not stated in the problem! But can be revealed by choosing an interesting test case).

TEST:
Submitted in-place variant via leetcode, and tested both approaches locally with the included simple test fixture.