Added 1720 Decode XORed array x2
PROBLEM: There is a hidden integer array arr that consists of n non-negative integers.
It was encoded into another integer array encoded of length n - 1, such that encoded[i] = arr[i] XOR arr[i + 1]. For example, if arr = [1,0,2,1], then encoded = [1,2,3].
You are given the encoded array. You are also given an integer first, that is the first element of arr, i.e. arr[0].
Return the original array arr. It can be proved that the answer exists and is unique.
https://leetcode.com/problems/decode-xored-array/

SOLUTION: Iterative approach that allocates a new vec, stores the first decoded value in the vec, and iterates through the encoded list, XORing each item with the running result, then adding the result to the new vec. The map approach similarly starts a new iterator with the first decoded value, then chains an iterator created by mapping each value of the encoded list to the running result of first XOR with the existing value, and returning the collected iterator.

TEST:
Submitted both approaches through leetcode:
https://leetcode.com/submissions/detail/531710024/
https://leetcode.com/submissions/detail/531709872/
and run locally using the included test fixture.