Added (not working) 1365 smaller numbers memoized
PROBLEM: https://leetcode.com/problems/how-many-numbers-are-smaller-than-the-current-number/
Attempting to use closures and Vec<Option<i32>> to memoize results. This is of dubious code improvement, but attempted for the exercise. This is currently blocked on borrow checker - writing to the memoization and reading from it are introducing conflicts that I'm not sure how to idiomatically resolve.

SOLUTION: smaller_numbers_memo_iterative sorts, and creates a list of options that may contain none. It maps the sorted list to memoize the results, checking if they've already been calculated, then maps through the original list the entries in the memo and returns.

TEST:
Code for this individual function does not build. TODO: Resolve borrow conflicts.