Added naive 91_decode_ways [FAILING]
PROBLEM: A message containing letters from A-Z can be encoded into numbers using the following mapping:
'A' -> "1"
'B' -> "2"
...
'Z' -> "26"
To decode an encoded message, all the digits must be grouped then mapped back into letters using the reverse of the mapping above (there may be multiple ways). For example, "11106" can be mapped into:
    "AAJF" with the grouping (1 1 10 6)
    "KJF" with the grouping (11 10 6)
Note that the grouping (1 11 06) is invalid because "06" cannot be mapped into 'F' since "6" is different from "06".
Given a string s containing only digits, return the number of ways to decode it.
The answer is guaranteed to fit in a 32-bit integer.

SOLUTION: Attempt to solve in one pass using iterators and match with windows of 3, handling the remaining 2 characters separately. This naive approach is ill-suited because of its inability to look ahead and branch, as evident in the test case "1123" which has 5 ways to decode, but iterators cannot reveal without tracking state.

TODO: Use a memoized / cached recursive approach

TEST:
https://leetcode.com/submissions/detail/540925228/
157 / 269 test cases passed.