Added two_sum_1 x2
PROBLEM: Given an array of integers nums and an integer target, return indices of the two numbers such that they add up to target.
You may assume that each input would have exactly one solution, and you may not use the same element twice.
You can return the answer in any order.

SOLUTION: Two solutions using iterators: the first maps from the front of the 'array' while finding from the back of the 'array' the element that adds to the target, then maps only if the indices are not equal. The second solution again uses map, but uses a position iterator on the subslice starting with the next index, and skips all mapped empty vectors, using iter.next() to return the first vector with valid entry.
28 ms and 12 ms runtimes respectively - this can come down (probably by eliminating position and/or find) if revisiting this problem.

TODO: Eliminate position and/or find, use another iterator approach that is performant!

TEST:
included test fixture with few test cases, and successful submissions:
https://leetcode.com/submissions/detail/540953246/
https://leetcode.com/submissions/detail/535137950/