Added 36 Valid Sudoku
PROBLEM: Determine if a 9 x 9 Sudoku board is valid. Only the filled cells need to be validated according to the following rules:
    Each row must contain the digits 1-9 without repetition.
    Each column must contain the digits 1-9 without repetition.
    Each of the nine 3 x 3 sub-boxes of the grid must contain the digits 1-9 without repetition.
Note:
    A Sudoku board (partially filled) could be valid but is not necessarily solvable.
    Only the filled cells need to be validated according to the mentioned rules.
Constraints:
    board.length == 9
    board[i].length == 9
    board[i][j] is a digit or '.'.
https://leetcode.com/problems/valid-sudoku/

SOLUTION: Two interesting notes: First, there is no mention of input inclusion or exclusion of digit 0, which could produce important input validation cases. I created a test case with 0 on the board, and their expected solution algorithm returned true, so I gather their intent is that boards will only have "1-9." characters, despite the loosely-worded constraint. Second, the input size is perfectly constrained for this style of problem, so we have much greater freedom in style of solution, since there is no concern for scaling runtimes with changing input sizes!
This included solution is a third revision of similar approaches: scan the board row by row, filtering out ".", and saving the board locations of each digit in a hash table (implemented in rust as a HashMap with vec values containing tuple pairs representing the points). Following this, iterate through each digit in the hash table with vec length > 1 (if there's only one occurence of a digit, it's valid). Remove a point from the vector and check against all other points to see if they are in the same row, or column, or box (both x and y integer divided by 3 equal for two points). If they are, they conflict and return invalid, otherwise keep checking until finished. If no conflicts found, the board is valid - return true.

TEST:
Tested via submission to leetcode:
https://leetcode.com/submissions/detail/541702681/
Earlier successful iterations, using a hash table constructed from Vectors, and a more demonstrative point-comparison function:
https://leetcode.com/submissions/detail/541693600/
https://leetcode.com/submissions/detail/541691208/