1630 arithmetic-subarrays solved + random test gen
Randomly generated 10 test cases using a manual test case generation function. This should be moved to a separate module, with data structures to group constraints and ranges based on the problem description and parameters.

The test case generation simply prints out the test cases in the same format as the leetcode source, when using `cargo run` you can redirect the output to a text file and copy from there to the leetcode ui.

The manual test harness does not yet read test cases from file: that's still a WIP.

For the problem solution, we have function `is_arithmetic(range)` which takes a slice from the nums vec, that calculates the difference of adjacent numbers in the slice (right minus left), and then returns true if all those differences are equal. Note the fold iterator adaptor starts special case None, and the rest of the diffs are handled in Option to avoid magic numbers.

Because the problem states we can rearrange any of the ranges, the most likely solutin is to sort the range, as if it is possible for the range to be an arithmetic subarray, and all differences will be equal, this will be reflected in a nondecreasing sequence. If a range isn't arithmetic, then we can attempt to sort the range to see if that sorted range is arithmetic, using `sort_range(range)`

Finally, to iterate through all the range start/end pairs contained in `l` and `r`, we zip them together (they are required to be same length, `m`), map the subsequence bounded by the range from nums, then run the `is_arithmetic` check on the range and if that fails, then on the sorted range, returning the `bool` result for each range pair, collecting them together and returning the collection.

10ms 83.33%
2.11 MB 83.33%