1727 largest-submatrix-with-rearrangements +soln post
https://leetcode.com/problems/largest-submatrix-with-rearrangements/solutions/4333497/26ms-7-97-mb-single-pass-scan-last-row-accumulator-o-n-x-m/?envType=daily-question&envId=2023-11-26

Intuition

We begin with the idea that we iterate through the entire matrix and add up (accumulate) vertical sequences of 1s. The resulting accumulated matrix will have cell values that represent the maximum height from that cell.

After that calculation, we can sort the rows in descending (nonincreasing) order, which will represent rearranging the columns optimally, as we have alread encoded the relationship to the other rows with the previous accumulation step.

Then we can iterate through the row and map the cell value (height) * the index + 1 (base) to get the max area from that cell as bottom right. This works because we know all the cells to the left are not less height, so the base will be valid all the way to the left. After mapping the max area for the cell, we can take the max of the row, and then take the max of all rows.
Approach

The solution can be optimized significantly by not requiring an additional pass through the matrix for the accumulation step. If we can store state of the last row as we are passing through, then we can calculate the accumulation for the current row and store it for the next pass in a stateful scan.

In the same pass, we can sort this accumulation (after we stored state for the next pass - the stored last_row needs to be in original order), then iterate through mapping the area formula and mapping just the max for the entire row.

This then requires finite n x m steps, or a single matrix pass while storing one row of state, and calculating the max along the way.
Complexity

    Time complexity:
    O(nxm)O(n x m)O(nxm)

    Space complexity:
    O(n)O(n)O(n)
    We only have to store 1-2 rows of state at any given time - including passing mutable clones to functions to perform the sort and max area. This is if we do not include the passed matrix as stored space (This algorithm can operate by reference on the input, although leetcode may not provide that option in our function signature)

---

Also re-wrote automatic test generation code and input [[i32; _]; _] to Vec<Vec<i32>> basic parsing.

TODO: Solve this programatically after detecting input parameters